refactor: split code into a module
